refactored cron
