+++ composer.json
@@ -9,7 +9,7 @@
         "contao/core-bundle": "^4.13",
         "terminal42/notification_center": "^1.4.1",
         "heimrichhannot/contao-formhybrid": "^3.2",
-        "heimrichhannot/contao-utils-bundle": "^2.19.0",
+        "heimrichhannot/contao-utils-bundle": "^2.235.1",
         "symfony/dependency-injection": "^5.4 || ^6.0"
     },
     "authors": [
