+++ config/services.yml
@@ -4,7 +4,7 @@ services:
 
   HeimrichHannot\ContaoNewsAlertBundle\:
     resource: '../src/{Components,DataContainer,Migration}/*'
-    exclude: '../src/Components/{DcaAddon.php,NewsTopicCollection.php,PoorManCron.php}'
+    exclude: '../src/Components/{DcaAddon.php,NewsTopicCollection.php}'
     public: false
     autoconfigure: true
     autowire: true
@@ -14,12 +14,18 @@ services:
     autoconfigure: true
     autowire: true
 
+  HeimrichHannot\ContaoNewsAlertBundle\EventListener\ContaoCronListener:
+    autoconfigure: true
+    autowire: true
+
 # EventListener
 
   huh.newsalert.listener.newsposted:
     class: HeimrichHannot\ContaoNewsAlertBundle\EventListener\NewsPostedListener
     public: true
     arguments: ['@service_container']
+  HeimrichHannot\ContaoNewsAlertBundle\EventListener\NewsPostedListener: '@huh.newsalert.listener.newsposted'
+
   huh.newsalert.listener.callback.table.modules:
     class: HeimrichHannot\ContaoNewsAlertBundle\EventListener\ModulesTableCallbackListener
     arguments: ['@contao.framework']
